requirements.txt
appnope==0.1.2
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
decorator==5.0.9
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
idna==2.10
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
jieba3k==0.35.1
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
lxml==4.6.3
matplotlib-inline==0.1.2
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
scikit-learn==0.24.2
scipy==1.6.3
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
